requirements.txt
PyQt6>=6.7
PyQt6-WebEngine>=6.7
openai>=1.0
python-dotenv>=1.0
pydantic>=2.8
mss>=10.0
pillow>=11.0
rapidocr-onnxruntime>=1.4
onnxruntime>=1.23.2
numpy>=1.26
edge-tts>=7.0
pygame>=2.6
azure-cognitiveservices-speech>=1.42
live2d-py>=0.5.4
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
